Add DLLInjector class
This class can inject dll file to other process.

But, you should run program using this 'Release-x64' compile option in
x64. Maybe... 'Release-Win32' too.
And this class use NtCreateThreadEx function because of management
policy of session in windows vista or later. So, this class can not be
running when changing native api(NtCreateThreadEx).